backend/requirements.txt
# Backend dependencies for PropPulse

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# HTTP & scraping
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.2

# NLP
spacy==3.7.6
# After install, also run: python -m spacy download en_core_web_sm
bertopic==0.16.4
sentence-transformers==3.1.1

# LLM / RAG
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
groq==0.11.0

# Data handling
pandas==2.2.3

# Environment variables
python-dotenv==1.0.1

# CORS (included in FastAPI but listed for clarity)
# fastapi includes starlette which handles CORS
